+++ session.go
@@ -16,7 +16,7 @@ import (
 	"time"
 	"unicode"
 
-	"github.com/gocql/gocql/lru"
+	"github.com/gocql/gocql/internal/lru"
 )
 
 // Session is the interface used by users to interact with the database.
